Create automation model
